install: Move action installation into a separate package
