--- multi-repo-action/go.mod
@@ -1,20 +1,0 @@
-module github.com/ossf/scorecard-actions/cli
-
-go 1.17
-
-require (
-	github.com/google/go-github/v42 v42.0.0
-	github.com/migueleliasweb/go-github-mock v0.0.6
-	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8
-)
-
-require (
-	github.com/golang/protobuf v1.4.2 // indirect
-	github.com/google/go-github/v41 v41.0.0 // indirect
-	github.com/google/go-querystring v1.1.0 // indirect
-	github.com/gorilla/mux v1.8.0 // indirect
-	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
-	golang.org/x/net v0.0.0-20210226172049-e18ecbb05110 // indirect
-	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/protobuf v1.25.0 // indirect
-)
